removed unused files, organized everything, removed warnings
