+++ client/src/components/Home/Home.js
@@ -13,6 +13,7 @@ function Home() {
       return;
     }
     if (user) history.replace(process.env.PUBLIC_URL + "/dashboard");
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [user, loading]);
   return (
     <>
