+++ client/src/components/Register/Register.js
@@ -18,6 +18,7 @@ function Register() {
   useEffect(() => {
     if (loading) return;
     if (user) history.replace("/dashboard");
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [user, loading]);
   return (
     <>
